minor dependency resolver change.

+++ t/01_loading.t
@@ -2,10 +2,14 @@
 use Test::More qw"no_plan";
 use File::Path qw"remove_tree";
 
-ok(remove_tree('scripts'),
-   'Removed scripts/');
-ok(remove_tree('outputs'),
-   'Removed outputs/');
+if (-d 'scripts') {
+    ok(remove_tree('scripts'),
+       'Removed scripts/');
+}
+if (-d 'outputs') {
+    ok(remove_tree('outputs'),
+       'Removed outputs/');
+}
 
 use_ok(Bio::Adventure);
 my $cyoa = new Bio::Adventure;
